+++ config.py
@@ -25,7 +25,7 @@ def save_history(totallist):
         num = num + 1
     dic2["historytrack"] = list1
     db.history.insert(dic2, check_keys=False)
-    
+
 
 def change_pos(data):
     totallist = []
@@ -39,14 +39,12 @@ def change_pos(data):
         num = num + 1
     totallist[data["tag"]] = data["location"]
     save_history(totallist)
-        
 
-    
+
+
 
 def load_config():
     cursor = db.config.find()
     for result in cursor:
         result.pop("_id")
-    config = json.dumps(result)
-    print(result)
-    return config
+    return result
