+++ views.py
@@ -9,58 +9,87 @@ from __future__ import print_function
 import base64
 import json
 import os
+import pymongo
+import sys
 
-from flask import make_response
-from flask import redirect
+from flask import Flask, render_template
 from flask import request
-from flask import render_template
 from flask import send_from_directory
-from flask import url_for
 
 from . import app
-from . import orion_dir
-from . import cfg_path, pos_path
-from ..config import load_config
-from ..config import save_config
 
 
+# Type of image
 VALID_IMG_EXT = ["jpg", "png"]
+
+# MongoDB client URL and port
+DEFAULT_URL = "mongodb://localhost:27017"
+
+
+client = pymongo.MongoClient(DEFAULT_URL)
+db = client["orion"]
 
 
 ################################ STATIC ROUTES ################################
 
 @app.route("/")
 def index():
-
     return render_template("index.html")
+
 
 @app.route("/fonts/<path:path>")
 def static_fonts():
-
     return send_from_directory(os.path.join("static", "fonts"))
+
 
 ################################ DYNAMIC PAGES ################################
 
 @app.route("/getconf")
 def getconf():
+    cursor = db.config.find()
+    for result in cursor:
+        result.pop("_id")
+    tagname = json.dumps(result)
 
-    return json.dumps(load_config(path=cfg_path, ldall=False))
+    return tagname
+
 
 @app.route("/setconf", methods=["POST"])
 def setconf():
+    data = request.get_json()
+    db.config.save(data, check_keys=False)
 
-    save_config(json.loads(request.get_data()), cfg_path)
     return json.dumps({"status": 1})
+
 
 @app.route("/positions")
 def positions():
+    num = 0
+    num1 = 0
+    posdata = []
+    posdata1 = []
+    posdata2 = []
+    cursor = db.history.find()
 
-    with open(pos_path, "r") as f:
-        return f.read()
+    for result in cursor:
+        result.pop("_id")
+        historytrack = result["historytrack"]
+        print(historytrack)
+        while num < len(historytrack):
+            posdata.append(historytrack[num]['pos'][0])
+            posdata.append(historytrack[num]['pos'][1])
+            num = num + 1
+            posdata1.append(posdata)
+            posdata = []
+        posdata2.append(posdata1)
+        posdata1 = []
+        num =0
+        print(posdata2)
+    return json.dumps(posdata2)
+
 
 @app.route("/upload", methods=["POST"])
 def upload():
-
     img = request.files["file"]
 
     # ensure that file extension is valid
@@ -79,13 +108,45 @@ def upload():
 
     return json.dumps({"status": 1, "img": img_b64})
 
+
 @app.route("/history", methods=["POST"])
 def history():
-
     # TODO(fzliu): get this working
     timedata = request.get_data()
     print(timedata)
     raise NotImplementedError()
+
+
+
+@app.route("/history_track", methods=["POST"])
+def history_track():
+    timedata = request.get_data()
+    timedata = json.loads(timedata)
+    num = 0
+    num1 = 0
+    posdata = []
+    posdata1 = []
+    posdata2 = []
+    cursor = db.history.find()
+
+    for result in cursor:
+        result.pop("_id")
+        historytrack = result["historytrack"]
+        print(historytrack)
+        while num < len(historytrack):
+            posdata.append(historytrack[num]['pos'][0])
+            posdata.append(historytrack[num]['pos'][1])
+            num = num + 1
+            posdata1.append(posdata)
+            posdata = []
+        posdata2.append(posdata1)
+        posdata1 = []
+        num =0
+        print(posdata2)
+    return json.dumps(posdata2)  # history_track needs to change
+
+
+
 
 ################################# DEBUG PAGES #################################
 
@@ -97,18 +158,22 @@ DEBUG_ROOM_SIZE = 25
 DEBUG_TAG_NAMES = ["Tag" + str(n) for n in range(DEBUG_NUM_TAGS)]
 DEBUG_TAG_POS = np.random.random((DEBUG_NUM_TAGS, 2)) * DEBUG_ROOM_SIZE
 
+
 @app.route("/_getconf")
 def _config():
-
     config = {
         "num_tags": DEBUG_NUM_TAGS,
         "tag_names": DEBUG_TAG_NAMES
     }
     return json.dumps(config)
 
+
 @app.route("/_positions")
 def _positions():
-
     pos_data = DEBUG_TAG_POS + np.random.random((DEBUG_NUM_TAGS, 2))
     pos_repr = np.array(40 * pos_data, dtype=np.int32).tolist()
     return json.dumps(pos_repr)
+
+
+if __name__ == "__main__":
+    app.run("0.0.0.0", 8000, debug=True)
